chatbot-assistant-ai/requirements.txt
python-dotenv
fastapi
uvicorn
streamlit
requests
ollama
python-docx 
fpdf
langdetect
SpeechRecognition
pyaudio
pydub
regex
langdetect>=1.0.9
ffmpeg-python
torch 
openai-whisper
TTS
scipy
ffmpeg 
numpy
pymupdf 
pdf2image 
pytesseract 
pillow 